Update urdfdom/urdfdom_headers to upstream 0.4 version
We switched to our forks mainly for Windows support
and to work around https://github.com/ros/urdfdom_headers/issues/18 ,
but since then the fix of the rpy handling issue has been released
in version 0.4 .

Furthermore, iDynTree on Windows used by default its own vendored urdfdom,
that however is only used for the legacy KDL classes.

For this reason we update to urdfdom to the latest 0.4 version (to avoid
incompatibilites if a user in Xenial installs both the packaged version of urdfdom
and do not clean up his install folder). On macOS tipically the user
install Gazebo before installing the superbuild, and so they get
urdfdom 1.0.0 installed by OSRF homebrew.
Eventually we should remove altogether urdfdom from the superbuild and
always rely on the one provided by the system package manager, but
I am a bit afraid that there are around release of urdfdom < 0.4 that
are affected by  https://github.com/ros/urdfdom_headers/issues/18 .
Enforcing the urdfdom version to be >= 0.4 is probably the right way to go
in the long run.