Modified "Joint calibration" section
merged sections "check the joint calibration" and "calibration of the robot"
changed the path to the calibration files, since different systems may have different installation
added details on the IMU calibration. I remember that the head-calib on iCubNancy is different from iCubGenova04, so I differentiated between CAN and ETH icub heads, hoping that the information is accurate.